+++ kaithem/src/plugins/startup/BareSIPPlugin/__init__.py
@@ -5,7 +5,7 @@ import os
 import mako
 import time
 import threading
-import logging
+import logging 
 import weakref
 import base64
 import traceback