+++ kaithem/src/thirdparty/scullery/thirdparty/baresipy/__init__.py
@@ -5,8 +5,8 @@ import tempfile
 import logging
 import subprocess
 from threading import Thread
-from baresipy.utils import create_daemon
-from baresipy.utils.log import LOG
+from  .utils import create_daemon
+from .utils.log import LOG
 import os
 import shutil
 
@@ -36,11 +36,11 @@ class BareSIP(Thread):
         if tts:
             self.tts = tts
         else:
-			try:
-				from responsive_voice import ResponsiveVoice
-				self.tts = ResponsiveVoice(gender=ResponsiveVoice.MALE)
-			except ImportError:
-				logging.exception("No responsive_voice module, some features will not work")
+            try:
+                from responsive_voice import ResponsiveVoice
+                self.tts = ResponsiveVoice(gender=ResponsiveVoice.MALE)
+            except ImportError:
+                logging.exception("No responsive_voice module, some features will not work")
         self._login = "sip:{u}:{p}@{g}".format(u=self.user, p=self.pwd,
                                                g=self.gateway)
         self._prev_output = ""
@@ -194,7 +194,7 @@ class BareSIP(Thread):
         self.abort = True
 
     def send_dtmf(self, number):
-		from opentone import ToneGenerator
+        from opentone import ToneGenerator
         number = str(number)
         for n in number:
             if n not in "0123456789":
@@ -228,7 +228,7 @@ class BareSIP(Thread):
 
     @staticmethod
     def convert_audio(input_file, outfile=None):
-		from pydub import AudioSegment
+        from pydub import AudioSegment
         input_file = expanduser(input_file)
         sound = AudioSegment.from_file(input_file)
         sound += AudioSegment.silent(duration=500)
@@ -237,7 +237,7 @@ class BareSIP(Thread):
         while sound.duration_seconds < 3:
             sound += AudioSegment.silent(duration=500)
 
-		outfile = outfile or join(tempfile.gettempdir(), "pybaresip.wav")
+        outfile = outfile or join(tempfile.gettempdir(), "pybaresip.wav")
         sound = sound.set_frame_rate(48000)
         sound = sound.set_channels(2)
         sound.export(outfile, format="wav")